requirements.txt
PyYAML
python-telegram-bot
requests
